+++ src/ORM/Map.php
@@ -260,13 +260,48 @@ class Map implements ArrayAccess, Countable, IteratorAggregate
     #[\ReturnTypeWillChange]
     public function getIterator()
     {
-        return new Map_Iterator(
-            $this->list->getIterator(),
-            $this->keyField,
-            $this->valueField,
-            $this->firstItems,
-            $this->lastItems
-        );
+        $keyField = $this->keyField;
+        $valueField = $this->valueField;
+
+        foreach($this->firstItems as $k => $v) {
+            yield $k => $v;
+        }
+
+        foreach($this->list as $record) {
+            if(isset($this->firstItems[$record->$keyField])) {
+                continue;
+            }
+            if(isset($this->lastItems[$record->$keyField])) {
+                continue;
+            }
+            yield $this->extractValue($record, $this->keyField) => $this->extractValue($record, $this->valueField);
+        }
+
+        foreach($this->lastItems as $k => $v) {
+            yield $k => $v;
+        }
+    }
+
+    /**
+     * Extracts a value from an item in the list, where the item is either an
+     * object or array.
+     *
+     * @param  array|object $item
+     * @param  string $key
+     * @return mixed
+     */
+    protected function extractValue($item, $key)
+    {
+        if (is_object($item)) {
+            if (method_exists($item, 'hasMethod') && $item->hasMethod($key)) {
+                return $item->{$key}();
+            }
+            return $item->{$key};
+        } else {
+            if (array_key_exists($key, $item)) {
+                return $item[$key];
+            }
+        }
     }
 
     /**
